requirements.txt
google-api-python-client>=2.0.0
google-auth-oauthlib>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
playwright>=1.40.0
python-magic>=0.4.27
dnspython>=2.4.0
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart
chromadb
sentence-transformers
ollama
